fix(types): narrow logger types, add explicit casts, add logging to 15+ modules
Phase 3b: Narrow logger type annotations from Logger | StructuredLogger to
StructuredLogger in component_factory, verification modules, stockfish_adapter.
Phase 3c: Add explicit casts for no-any-return errors in settings.py,
chess/state.py, policies.py, s3_client.py, base.py, and 15+ other files.
Phase 4: Add logging.getLogger(__name__) to auth.py, hrm_agent.py, trm_agent.py,
metrics.py, policy_network.py, value_network.py, policy_value_net.py,
data_generator.py, replay_buffer.py, neuro_symbolic modules, and graph.py.

Mypy errors reduced from 336 to 215 (121 errors fixed).

+++ src/observability/profiling.py
@@ -276,7 +276,7 @@ class AsyncProfiler:
         if session_id and session_id in self._sessions:
             self._sessions[session_id].cpu_samples.append(cpu_percent)
 
-        return cpu_percent
+        return float(cpu_percent)
 
     def add_marker(
         self,
@@ -360,13 +360,13 @@ class MemoryProfiler:
 
     def set_baseline(self) -> float:
         """Set current memory as baseline."""
-        self._baseline = self._process.memory_info().rss / (1024 * 1024)
+        self._baseline = float(self._process.memory_info().rss / (1024 * 1024))
         self.logger.info(f"Memory baseline set: {self._baseline:.2f} MB")
         return self._baseline
 
     def get_current(self) -> float:
         """Get current memory usage in MB."""
-        return self._process.memory_info().rss / (1024 * 1024)
+        return float(self._process.memory_info().rss / (1024 * 1024))
 
     def get_delta(self) -> float:
         """Get memory change from baseline."""
